add blake2 state success tests
