--- core/crypto/blake2/tests/compression.rs
@@ -1,76 +1,0 @@
-use blake2;
-
-// https://tools.ietf.org/html/rfc7693#appendix-A
-#[test]
-fn blake2b_f_function() {
-    let rounds = 12;
-    let h: [u64; 8] = [
-        0x6a09e667f2bdc948,
-        0xbb67ae8584caa73b,
-        0x3c6ef372fe94f82b,
-        0xa54ff53a5f1d36f1,
-        0x510e527fade682d1,
-        0x9b05688c2b3e6c1f,
-        0x1f83d9abfb41bd6b,
-        0x5be0cd19137e2179,
-    ];
-    let m: [u64; 16] = [
-        0x0000000000636261,
-        0x0000000000000000,
-        0x0000000000000000,
-        0x0000000000000000,
-        0x0000000000000000,
-        0x0000000000000000,
-        0x0000000000000000,
-        0x0000000000000000,
-        0x0000000000000000,
-        0x0000000000000000,
-        0x0000000000000000,
-        0x0000000000000000,
-        0x0000000000000000,
-        0x0000000000000000,
-        0x0000000000000000,
-        0x0000000000000000,
-    ];
-    let t: [u64; 2] = [3, 0];
-    let f_bool = true;
-
-    let output: [u8; 64] = [
-        0xba, 0x80, 0xa5, 0x3f, 0x98, 0x1c, 0x4d, 0x0d, 0x6a, 0x27, 0x97, 0xb6, 0x9f, 0x12, 0xf6,
-        0xe9, 0x4c, 0x21, 0x2f, 0x14, 0x68, 0x5a, 0xc4, 0xb7, 0x4b, 0x12, 0xbb, 0x6f, 0xdb, 0xff,
-        0xa2, 0xd1, 0x7d, 0x87, 0xc5, 0x39, 0x2a, 0xab, 0x79, 0x2d, 0xc2, 0x52, 0xd5, 0xde, 0x45,
-        0x33, 0xcc, 0x95, 0x18, 0xd3, 0x8a, 0xa8, 0xdb, 0xf1, 0x92, 0x5a, 0xb9, 0x23, 0x86, 0xed,
-        0xd4, 0x0, 0x99, 0x23,
-    ];
-
-    let res = blake2::blake2b_f(rounds, h, m, t, f_bool);
-
-    assert_eq!(res.as_slice(), output);
-}
-
-// https://tools.ietf.org/html/rfc7693#appendix-A
-#[test]
-fn blake2s_f_function() {
-    let rounds = 10;
-    let h: [u32; 8] = [
-        0x6B08E647, 0xBB67AE85, 0x3C6EF372, 0xA54FF53A, 0x510E527F, 0x9B05688C, 0x1F83D9AB,
-        0x5BE0CD19,
-    ];
-    let m: [u32; 16] = [
-        0x00636261, 0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000,
-        0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000,
-        0x00000000, 0x00000000,
-    ];
-    let t: [u32; 2] = [3, 0];
-    let f_bool = true;
-
-    let output: &[u8; 32] = &[
-        0x50, 0x8c, 0x5e, 0x8c, 0x32, 0x7c, 0x14, 0xe2, 0xe1, 0xa7, 0x2b, 0xa3, 0x4e, 0xeb, 0x45,
-        0x2f, 0x37, 0x45, 0x8b, 0x20, 0x9e, 0xd6, 0x3a, 0x29, 0x4d, 0x99, 0x9b, 0x4c, 0x86, 0x67,
-        0x59, 0x82,
-    ];
-
-    let res = blake2::blake2s_f(rounds, h, m, t, f_bool);
-
-    assert_eq!(res.as_slice(), output);
-}
